First Commit
This is my first commit for PS2

+++ PS2/problem2.cpp
@@ -47,7 +47,7 @@ using namespace std;
 int main() {
     for (int i = 0; i < 3; i++) {
         for (int j = 3; j > 0; j--) {
-            cout << pow(i, j) << endl;
+            cout << pow(static_cast<double>(i), j) << endl;
         }
     }
 }
